Delete command for tests implemented [alpha-version, improvement needed]

+++ src/main/java/com/javacourse/test/answer/AnswerDAO.java
@@ -8,7 +8,7 @@ import java.util.List;
 /**
  * Basic interface used for implementing DAOFactory for switching between databases easily
  */
-public interface AnswerDAO extends AbstractDAO<Integer, Answer> {
+public interface AnswerDAO extends AbstractDAO<Long, Answer> {
     List<Answer> findCorrectAnswersByTaskId(long task_id) throws UnsuccessfulQueryException;
     List<Answer> findPossibleAnswersByTaskId(long task_id) throws UnsuccessfulQueryException;
 }
